% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_metrics}
\alias{build_metrics}
\title{Build metrics array for API}
\usage{
build_metrics(
  metric_fob = TRUE,
  metric_kg = TRUE,
  metric_statistic = FALSE,
  metric_freight = FALSE,
  metric_insurance = FALSE,
  metric_cif = FALSE
)
}
\description{
Build metrics array for API
}
\keyword{internal}
