% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comex_close}
\alias{comex_close}
\title{Close browser session}
\usage{
comex_close()
}
\value{
Invisibly returns TRUE
}
\description{
Closes the browser session and frees resources.
}
\examples{
\dontrun{
comex_close()
}

}
