% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{comex_country_detail}
\alias{comex_country_detail}
\title{Get country details}
\usage{
comex_country_detail(id, verbose = FALSE)
}
\arguments{
\item{id}{Country code (e.g. \code{105} for Brazil).}

\item{verbose}{Logical. Default: \code{FALSE}.}
}
\value{
A list with country details.
}
\description{
Returns details for a specific country by its code.
}
\examples{
\dontrun{
comex_country_detail(105)
}

}
