% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{comex_customs_unit_detail}
\alias{comex_customs_unit_detail}
\title{Get customs unit details}
\usage{
comex_customs_unit_detail(urf_id, language = "en", verbose = FALSE)
}
\arguments{
\item{urf_id}{Customs unit code}

\item{language}{Language: "pt", "en", or "es". Default: "en"}

\item{verbose}{Logical. If TRUE, display progress messages. Default: FALSE}
}
\value{
A list with customs unit details
}
\description{
Returns details for a specific customs unit.
}
\examples{
\dontrun{
comex_customs_unit_detail(817600)
}

}
