% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comexstat_download_2.R
\name{comex_download}
\alias{comex_download}
\title{Download Comexstat Data from MDIC (Brazilian Ministry of Development, Industry, Commerce, and Services)}
\usage{
comex_download(
  years = 2024,
  directions = c("imp", "exp"),
  types = c("hs4", "ncm"),
  download_aux = TRUE,
  clean_aux = TRUE,
  cache = TRUE,
  .progress = TRUE,
  n_tries = 30,
  check = FALSE,
  ...
)
}
\arguments{
\item{years}{A numeric vector or integer specifying the years for which data should be downloaded. Defaults to the current year (\code{2024}).}

\item{directions}{A character vector specifying the directions of trade: 'imp' (imports) and/or 'exp' (exports). Defaults to both.}

\item{types}{A character vector specifying the types of data: 'ncm' (Nomenclatura Comum do Mercosul) and/or 'hs4' (Harmonized System 4-digit). Defaults to both.}

\item{download_aux}{A logical value indicating whether to download auxiliary data tables (e.g., URF, VIA, country codes). Defaults to \code{TRUE}.}

\item{clean_aux}{A logical value indicating whether to delete existing auxiliary data files before downloading new ones. Defaults to \code{TRUE}.}

\item{cache}{A logical value indicating whether to use cached files if they exist. Defaults to \code{TRUE}.}

\item{.progress}{A logical value indicating whether to display a progress bar during downloads. Defaults to \code{TRUE}.}

\item{n_tries}{The maximum number of download attempts before giving up. Defaults to 30.}

\item{check}{A logical value indicating whether to perform data validation checks after downloading. Defaults to \code{FALSE}.}

\item{...}{Additional arguments to be passed to \code{curl::multi_download}, such as \code{headers}, \code{handle}, etc.}
}
\value{
No direct return value. The function downloads the requested files to the specified directories.
}
\description{
This function downloads Comexstat (Brazilian trade statistics) data from the
MDIC website for specified years, directions (imports/exports), and
types (NCM/HS4). It can also optionally download and manage auxiliary data tables.
}
\details{
This function constructs the URLs for the specified Comexstat data files based on the \code{years}, \code{directions}, and \code{types} arguments.
It then creates the necessary directories to store the data.

Downloads are performed using \code{curl::multi_download}, with retry logic if there are initial failures. If \code{download_aux} is \code{TRUE},
additional auxiliary data tables are downloaded as well.

If \code{check} is \code{TRUE}, a basic data validation check is performed using the internal \code{comex_check} function.
}
\examples{
\dontrun{
# Download all data for 2023 and 2024
comex_download(years = 2023:2024)

# Download only import data (NCM) for 2024
comex_download(years = 2024, directions = 'imp', types = 'ncm')
}

}
