% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_products.R
\name{comex_nbm_detail}
\alias{comex_nbm_detail}
\title{Get NBM code details}
\usage{
comex_nbm_detail(nbm_code, verbose = FALSE)
}
\arguments{
\item{nbm_code}{NBM code (e.g. \code{"2924101100"}).}

\item{verbose}{Logical. Default: \code{FALSE}.}
}
\value{
A list with NBM details.
}
\description{
Returns details for a specific NBM code.
}
\examples{
\dontrun{
comex_nbm_detail("2924101100")
}

}
