% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_products.R
\name{comex_ncm}
\alias{comex_ncm}
\title{Get NCM (Mercosur Common Nomenclature) table}
\usage{
comex_ncm(language = "en", verbose = FALSE)
}
\arguments{
\item{language}{Language: "pt", "en", or "es". Default: "en"}

\item{verbose}{Logical. If TRUE, display progress messages. Default: FALSE}
}
\value{
A tibble with NCM codes and descriptions
}
\description{
Returns the NCM codes table with descriptions. NCM is the product
classification used by Mercosur countries, based on the Harmonized
System (HS).
}
\examples{
\dontrun{
comex_ncm()
}

}
