% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{comex_transport_mode_detail}
\alias{comex_transport_mode_detail}
\title{Get transport mode details}
\usage{
comex_transport_mode_detail(mode_id, verbose = FALSE)
}
\arguments{
\item{mode_id}{Transport mode code (e.g. \code{5} for maritime).}

\item{verbose}{Logical. Default: \code{FALSE}.}
}
\value{
A list with transport mode details.
}
\description{
Returns details for a specific transport mode.
}
\examples{
\dontrun{
comex_transport_mode_detail(5)
}

}
