% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{comex_transport_modes}
\alias{comex_transport_modes}
\title{Get transport modes table}
\usage{
comex_transport_modes(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. Default: \code{FALSE}.}
}
\value{
A data.frame with transport mode codes and names.
}
\description{
Returns the transport modes table with codes and names.
}
\examples{
\dontrun{
comex_transport_modes()
}

}
