% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comexstat2.R
\name{comexstat}
\alias{comexstat}
\title{Note that arrow only accepts a subset of the dplyr functions. The remaining tables are read from the original csv files.}
\usage{
comexstat(table, ...)
}
\description{
Note that arrow only accepts a subset of the dplyr functions. The remaining tables are read from the original csv files.
}
\examples{
comexstat2_download()
comexstat2("ncm")|>head()
}
