% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comexstat2.R
\name{comexstat_deflated}
\alias{comexstat_deflated}
\title{Deflate comexstat series}
\usage{
comexstat_deflated(
  data = comexstat_ncm(),
  basedate = NULL,
  deflators = get_deflators(na_omit = TRUE)
)
}
\arguments{
\item{data}{trade data to deflate}

\item{basedate}{base date to deflate to.}

\item{deflators}{data.frame with deflators}
}
\value{
deflated data.
#' @examples
library(dplyr)
total_trade <- comexstat_ncm()\%>\%group_by(year, date)\%>\%summarise(fob_usd=sum(fob_usd))\%>\%comexstat_deflated()\%>\%group_by(year)\%>\%summarise(fob_usd=sum(fob_usd),fob_usd_constant=sum(fob_usd_constant))\%>\%collect()
total_trade\%>\%select(year, fob_usd, fob_usd_constant)\%>\%arrange(desc(fob_usd_constant))
}
\description{
Deflate comexstat series
}
