% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{comexstat_m}
\alias{comexstat_m}
\title{Create id (co_ano_mes_m) by m months}
\usage{
comexstat_m(m = 12, data = comextat())
}
\arguments{
\item{m}{number of months}
}
\value{
tibble with id variables: co_ano_mes_m, co_pais, fluxo, co_ncm; and sums of qt_estat, vl_cif, vl_fob
}
\description{
Create id (co_ano_mes_m) by m months
}
\details{
Since it outputs every possible combination of co_ano_mes_m, co_pais, fluxo and co_ncm, it can be very large, and time consuming.
}
