% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comexstat2.R
\name{comexstat_ncm}
\alias{comexstat_ncm}
\title{Open ComexStat NCM Trade Dataset}
\usage{
comexstat_ncm()
}
\value{
An Arrow Dataset containing combined import and export NCM trade data, with an additional \code{direction} column indicating "exp" (export) or "imp" (import). The dataset has the following columns:
\itemize{
\item \code{year}: Year (integer)
\item \code{month}: Month (integer)
\item \code{ncm}: 8-digit NCM product code (string)
\item \code{unit_code}: Unit of measurement code (integer)
\item \code{country_code}: Country code (integer)
\item \code{state_abb}: Brazilian state abbreviation (string)
\item \code{transp_mode_code}: Transportation mode code (integer)
\item \code{urf_code}: Customs clearance unit code (integer)
\item \code{qt_stat}: Statistical quantity (integer64)
\item \code{kg_net}: Net weight in kilograms (integer64)
\item \code{fob_usd}: FOB value in US dollars (integer64)
\item \code{freight_usd}: Freight value in US dollars (integer64, only for imports)
\item \code{insurance_usd}: Insurance value in US dollars (integer64, only for imports)
\item \code{direction}: Trade direction, either "exp" (export) or "imp" (import) (derived)
}
}
\description{
This function opens the ComexStat NCM (8-digit product code) trade data as an Arrow Dataset. The data is assumed to be located in the  ComexStat data directory and to have been downloaded using the \code{comexstat_download()} function.
}
\details{
This function assumes that the NCM data files are named according to the following pattern:
\itemize{
\item Imports: "imp_\link{year}.csv" (e.g., "imp_2023.csv")
\item Exports: "exp_\link{year}.csv" (e.g., "exp_2023.csv")
}

It infers the trade direction based on the presence or absence of the \code{freight_usd} column (present only in imports).
}
\examples{
\dontrun{
library(dplyr)
# Open the ComexStat NCM dataset:
ncm_dataset <- comexstat_ncm()

# Explore the dataset:
print(ncm_dataset)
ncm_dataset|>group_by(year, direction)|>summarise(fob_usd=sum(fob_usd), freight_usd=sum(freight_usd))|>mutate(p=freight_usd/fob_usd)\%>\%collect()
comexstat_ncm()\%>\%group_by(country_code)\%>\%summarise(fob_usd=sum(fob_usd, na.rm=TRUE), freight_usd=sum(freight_usd))\%>\%mutate(p=freight_usd/fob_usd)\%>\%arrange(desc(p))\%>\%collect()
}

}
