% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{comexstat_raw}
\alias{comexstat_raw}
\title{Reads files with comexdata in the cache directory}
\usage{
comexstat_raw()
}
\value{
arrow::open_dataset
}
\description{
Reads files with comexdata in the cache directory
}
\details{
The downloaded files are one for the exports and other for imports, with data for the entire period available (1997-).

This function reads those files using arrow. It is used by comexstat_rewrite.
}
