% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{comexstat_rewrite}
\alias{comexstat_rewrite}
\title{Rewrites the data read from cache directory into partitioned files.}
\usage{
comexstat_rewrite()
}
\description{
Rewrites the data read from cache directory into partitioned files.
}
\details{
The downloaded files are one for the exports and other for imports, with data for the entire period available (1997-).

This function reads those files using arrow by calling function comexstat_raw. It then writes to the data directory the partitioned files.
}
