% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_comex}
\alias{download_comex}
\title{Internal function that downloads comextat data}
\usage{
download_comex(filenames, outdir = ddircomex, ...)
}
\arguments{
\item{filenames}{file to download, all for everything}

\item{outdir}{where to download the files to}

\item{...}{arguments sent to the base download.file function}
}
\value{
paths to the files downloaded
}
\description{
Internal function that downloads comextat data
}
