% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_deflators}
\alias{get_deflators}
\title{Download and processes deflators (CPI/USA, IPCA/Brazil, Exchange rate BRL/USE)}
\usage{
get_deflators(updated = Sys.Date())
}
\arguments{
\item{updated}{}
}
\description{
Download and processes deflators (CPI/USA, IPCA/Brazil, Exchange rate BRL/USE)
}
