% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{read_comex}
\alias{read_comex}
\title{Reads comexstat files from the specified directory}
\usage{
read_comex(name, dir = ddircomex, extension = ".csv")
}
\arguments{
\item{name}{name of the comexstat files, without extension}

\item{dir}{directory where the files are. defaults to the data directory specified in the package using rappdirs::user_data_dir}

\item{extension}{file extension to be read, defaults to .csv}
}
\value{
data.frame/tibble
}
\description{
Reads comexstat files from the specified directory
}
