% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{ym}
\alias{ym}
\title{Create a data frame with a co_ano_mes_m column with an id every m months}
\usage{
ym(m = 12, data = comexstat())
}
\arguments{
\item{m}{number of months to aggregate}

\item{data}{data frame or outpur of comexstat with co_ano and co_mes columns}
}
\value{
tibble with columns: co_ano, co_mes, co_ano_mes_m
}
\description{
Create a data frame with a co_ano_mes_m column with an id every m months
}
\examples{
ym(m=6, comexstat()|>filter(co_ano>=2021))
}
