% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{comex_available_years}
\alias{comex_available_years}
\title{Get available years for queries}
\usage{
comex_available_years(type = "general", verbose = FALSE)
}
\arguments{
\item{type}{Data type: "general", "city", or "historical"}

\item{verbose}{Logical. If TRUE, display progress messages. Default: FALSE}
}
\value{
A list with minimum and maximum available years
}
\description{
Returns the first and last years available for queries in the API.
}
\examples{
\dontrun{
comex_available_years()
comex_available_years(type = "city")
comex_available_years(type = "historical")
}

}
