% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comexstat2.R
\name{comex_check}
\alias{comex_check}
\title{Check Consistency Between Comex Data and Conference Files}
\usage{
comex_check(years = NULL, directions = NULL, type = "ncm")
}
\arguments{
\item{years}{A numeric vector specifying the years to check. If \code{NULL} (default), all available years are checked.}

\item{directions}{A character vector specifying the directions of trade to check: "imp" (imports) and/or "exp" (exports). If \code{NULL} (default), both directions are checked.}

\item{type}{A character string indicating the type of Comex data: "ncm" (Nomenclatura Comum do Mercosul) or "hs4" (Harmonized System 4-digit). Defaults to "ncm".}
}
\value{
No direct return value. The function prints a message indicating whether the check passed ("All clear!") or if there are mismatches between the data and conference files. If mismatches are found, a table displaying the discrepancies is printed and an error is raised.
}
\description{
This function verifies the consistency between downloaded Comex data and the corresponding conference files provided by the official source (MDIC). It checks if the aggregated totals for specific columns match between the two sources.
}
\details{
This function performs the following checks:
\enumerate{
\item \strong{Input Validation:} Ensures that the \code{type} argument is valid ("ncm" or "hs4") and has a length of 1.
\item \strong{Data Retrieval:} Reads the relevant Comex data (based on the \code{type}) and the corresponding conference files.
\item \strong{Filtering:} Filters the data and conference files based on the specified \code{years} and \code{directions} if provided.
\item \strong{Aggregation:} Aggregates the Comex data by \code{year} and \code{direction} and calculates sums for relevant columns.
\item \strong{Comparison:} Compares the aggregated sums from the Comex data with the totals in the conference files.
\item \strong{Mismatch Reporting:} If any mismatches are found, a table highlighting the differences is printed to the console.
\item \strong{Error Handling:} If mismatches exist, the function raises an error to stop further analysis.
}
}
\examples{
\dontrun{
# Check consistency for all NCM data from 2020 onwards
comex_check(years = 2020:2023)

# Check consistency for only import data (NCM)
comex_check(directions = "imp")

# Check consistency for HS4 data in 2022
comex_check(years = 2022, type = "hs4")
}

}
