% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{comex_cities}
\alias{comex_cities}
\title{Get Brazilian cities table}
\usage{
comex_cities(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. Default: \code{FALSE}.}
}
\value{
A data.frame with city (IBGE) codes and names.
}
\description{
Returns the Brazilian cities table with codes and names.
}
\examples{
\dontrun{
comex_cities()
}

}
