% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{comex_city_detail}
\alias{comex_city_detail}
\title{Get city details}
\usage{
comex_city_detail(city_id, verbose = FALSE)
}
\arguments{
\item{city_id}{IBGE city code (e.g. \code{5300050}).}

\item{verbose}{Logical. Default: \code{FALSE}.}
}
\value{
A list with city details.
}
\description{
Returns details for a specific Brazilian city.
}
\examples{
\dontrun{
comex_city_detail(5300050)
}

}
