% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{comex_country_detail}
\alias{comex_country_detail}
\title{Get country details}
\usage{
comex_country_detail(id, language = "en", verbose = FALSE)
}
\arguments{
\item{id}{Country code}

\item{language}{Language: "pt", "en", or "es". Default: "en"}

\item{verbose}{Logical. If TRUE, display progress messages. Default: FALSE}
}
\value{
A list with country details
}
\description{
Returns details for a specific country.
}
\examples{
\dontrun{
comex_country_detail(105)
}

}
