% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{comex_customs_unit_detail}
\alias{comex_customs_unit_detail}
\title{Get customs unit details}
\usage{
comex_customs_unit_detail(urf_id, verbose = FALSE)
}
\arguments{
\item{urf_id}{Customs unit code (e.g. \code{8110000}).}

\item{verbose}{Logical. Default: \code{FALSE}.}
}
\value{
A list with customs unit details.
}
\description{
Returns details for a specific customs unit (URF).
}
\examples{
\dontrun{
comex_customs_unit_detail(8110000)
}

}
