% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{comex_details}
\alias{comex_details}
\title{Get list of available details}
\usage{
comex_details(type = "general", language = "en", verbose = FALSE)
}
\arguments{
\item{type}{Data type: "general", "city", or "historical"}

\item{language}{Language: "pt", "en", or "es". Default: "en"}

\item{verbose}{Logical. If TRUE, display progress messages. Default: FALSE}
}
\value{
A tibble with available details
}
\description{
Returns the list of available detail/grouping fields.
}
\examples{
\dontrun{
comex_details()
comex_details(type = "city")
comex_details(type = "historical")
}

}
