% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{comex_filters}
\alias{comex_filters}
\title{Get list of available filters}
\usage{
comex_filters(type = "general", language = "en", verbose = FALSE)
}
\arguments{
\item{type}{Data type: "general", "city", or "historical"}

\item{language}{Language: "pt", "en", or "es". Default: "en"}

\item{verbose}{Logical. If TRUE, display progress messages. Default: FALSE}
}
\value{
A tibble with available filters
}
\description{
Returns the list of available filters for API queries.
}
\examples{
\dontrun{
comex_filters()
comex_filters(type = "city")
comex_filters(type = "historical")
}

}
