% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comexstat2.R
\name{comex_hs4}
\alias{comex_hs4}
\title{Open ComexStat HS4 Trade Dataset}
\usage{
comex_hs4()
}
\value{
An Arrow Dataset containing combined import and export HS4 trade data. The dataset has the following columns:
\itemize{
\item \code{year}: Year (integer)
\item \code{month}: Month (integer)
\item \code{hs4}: 4-digit HS product code (integer)
\item \code{country_code}: Country code (integer)
\item \code{state}: Brazilian state abbreviation (string)
\item \code{mun_code}: Municipality code (integer)
\item \code{kg_net}: Net weight in kilograms (integer64)
\item \code{fob_usd}: FOB value in US dollars (integer64)
\item \code{direction}: Trade direction, either 'exp' (export) or 'imp' (import)
}
}
\description{
This function opens the ComexStat HS4 trade data as an Arrow Dataset. The data is assumed to be located in the ComexStat data directory and to have been downloaded using the \code{comex_download()} function.
}
\details{
This function assumes that the HS4 data files are named according to the following pattern:
\itemize{
\item Imports: 'imp_\link{year}_mun.csv' (e.g., 'imp_2023_mun.csv')
\item Exports: 'exp_\link{year}_mun.csv' (e.g., 'exp_2023_mun.csv')
}
}
\examples{
\dontrun{
# Open the ComexStat HS4 dataset:
hs4_dataset <- comexstat_hs4()

# Explore the dataset:
print(hs4_dataset)
}

}
