% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{comex_last_update}
\alias{comex_last_update}
\title{Get last data update date}
\usage{
comex_last_update(type = "general", verbose = FALSE)
}
\arguments{
\item{type}{Data type: "general", "city", or "historical"}

\item{verbose}{Logical. If TRUE, display progress messages. Default: FALSE}
}
\value{
A list with last update information
}
\description{
Returns the date of the last data update in the API.
}
\examples{
\dontrun{
comex_last_update()
comex_last_update(type = "city")
comex_last_update(type = "historical")
}

}
