% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{comex_metrics}
\alias{comex_metrics}
\title{Get list of available metrics}
\usage{
comex_metrics(type = "general", language = "en", verbose = FALSE)
}
\arguments{
\item{type}{Data type: "general", "city", or "historical"}

\item{language}{Language: "pt", "en", or "es". Default: "en"}

\item{verbose}{Logical. If TRUE, display progress messages. Default: FALSE}
}
\value{
A tibble with available metrics
}
\description{
Returns the list of available metrics for API queries.
}
\examples{
\dontrun{
comex_metrics()
comex_metrics(type = "city")
comex_metrics(type = "historical")
}

}
