% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_products.R
\name{comex_nbm_detail}
\alias{comex_nbm_detail}
\title{Get NBM code details}
\usage{
comex_nbm_detail(nbm_code, language = "en", verbose = FALSE)
}
\arguments{
\item{nbm_code}{NBM code}

\item{language}{Language: "pt", "en", or "es". Default: "en"}

\item{verbose}{Logical. If TRUE, display progress messages. Default: FALSE}
}
\value{
A list with NBM details
}
\description{
Returns details for a specific NBM code.
}
\examples{
\dontrun{
comex_nbm_detail("00100101")
}

}
