% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_products.R
\name{comex_ncm_detail}
\alias{comex_ncm_detail}
\title{Get NCM code details}
\usage{
comex_ncm_detail(ncm_code, verbose = FALSE)
}
\arguments{
\item{ncm_code}{NCM code (8 digits, as character, e.g. \code{"02042200"}).}

\item{verbose}{Logical. Default: \code{FALSE}.}
}
\value{
A list with NCM details.
}
\description{
Returns details for a specific NCM code, including product description
and its HS classification hierarchy.
}
\examples{
\dontrun{
comex_ncm_detail("02042200")
}

}
