% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{comex_state_detail}
\alias{comex_state_detail}
\title{Get state details}
\usage{
comex_state_detail(uf_id, verbose = FALSE)
}
\arguments{
\item{uf_id}{State code (e.g. \code{26} for Pernambuco).}

\item{verbose}{Logical. Default: \code{FALSE}.}
}
\value{
A list with state details.
}
\description{
Returns details for a specific Brazilian state.
}
\examples{
\dontrun{
comex_state_detail(26)
}

}
