% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{comex_states}
\alias{comex_states}
\title{Get Brazilian states (UF) table}
\usage{
comex_states(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. Default: \code{FALSE}.}
}
\value{
A data.frame with state codes and names.
}
\description{
Returns the Brazilian states table with codes and names.
}
\examples{
\dontrun{
comex_states()
}

}
