% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comexr-package.R
\docType{package}
\name{comexr-package}
\alias{comexr}
\alias{comexr-package}
\title{comexr: R Client for the ComexStat API}
\description{
The \code{comexr} package provides a complete R interface to the ComexStat API
from the Brazilian Ministry of Development, Industry, Trade and Services
(MDIC).

The API provides Brazilian foreign trade data from 1997 (general data) and
from 1989 (historical data), along with auxiliary tables for product
nomenclatures, countries, economic classifications, and more.
}
\section{Query functions (POST)}{

These functions query aggregated export and import data with configurable
filters, details, and metrics:
\itemize{
\item \code{\link[=comex_query]{comex_query()}}: Query general foreign trade data
\item \code{\link[=comex_export]{comex_export()}}: Shortcut for export queries
\item \code{\link[=comex_import]{comex_import()}}: Shortcut for import queries
\item \code{\link[=comex_query_city]{comex_query_city()}}: Query city-level data
\item \code{\link[=comex_historical]{comex_historical()}}: Query historical data (1989-1996)
}
}

\section{API metadata (GET)}{

Functions to retrieve information about the API itself:
\itemize{
\item \code{\link[=comex_last_update]{comex_last_update()}}: Date of last data update
\item \code{\link[=comex_available_years]{comex_available_years()}}: Available years for queries
\item \code{\link[=comex_filters]{comex_filters()}}: Available filters
\item \code{\link[=comex_filter_values]{comex_filter_values()}}: Possible values for a filter
\item \code{\link[=comex_details]{comex_details()}}: Available detail/grouping fields
\item \code{\link[=comex_metrics]{comex_metrics()}}: Available metrics
}
}

\section{Auxiliary tables - Geography (GET)}{

\itemize{
\item \code{\link[=comex_countries]{comex_countries()}}, \code{\link[=comex_country_detail]{comex_country_detail()}}
\item \code{\link[=comex_blocs]{comex_blocs()}}
\item \code{\link[=comex_states]{comex_states()}}, \code{\link[=comex_state_detail]{comex_state_detail()}}
\item \code{\link[=comex_cities]{comex_cities()}}, \code{\link[=comex_city_detail]{comex_city_detail()}}
\item \code{\link[=comex_transport_modes]{comex_transport_modes()}}, \code{\link[=comex_transport_mode_detail]{comex_transport_mode_detail()}}
\item \code{\link[=comex_customs_units]{comex_customs_units()}}, \code{\link[=comex_customs_unit_detail]{comex_customs_unit_detail()}}
}
}

\section{Auxiliary tables - Products (GET)}{

\itemize{
\item \code{\link[=comex_ncm]{comex_ncm()}}, \code{\link[=comex_ncm_detail]{comex_ncm_detail()}}
\item \code{\link[=comex_nbm]{comex_nbm()}}, \code{\link[=comex_nbm_detail]{comex_nbm_detail()}}
\item \code{\link[=comex_hs]{comex_hs()}}
}
}

\section{Auxiliary tables - Classifications (GET)}{

\itemize{
\item \code{\link[=comex_cgce]{comex_cgce()}}
\item \code{\link[=comex_sitc]{comex_sitc()}}
\item \code{\link[=comex_isic]{comex_isic()}}
}
}

\section{Dependencies}{

The package uses \strong{only two dependencies}:
\itemize{
\item \code{httr2} for all HTTP requests (no headless browser needed)
\item \code{cli} for formatted console messages
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://strategicprojects.github.io/comexr/}
  \item \url{https://github.com/StrategicProjects/comexr}
  \item Report bugs at \url{https://github.com/StrategicProjects/comexr/issues}
}

}
\author{
\strong{Maintainer}: Andre Leite \email{leite@castlab.org}

Authors:
\itemize{
  \item Andre Leite \email{leite@castlab.org}
  \item Marcos Wasilew \email{marcos.wasilew@gmail.com}
  \item Hugo Vasconcelos \email{hugo.vasconcelos@ufpe.br}
  \item Carlos Amorim \email{carlos.agaf@ufpe.br}
  \item Diogo Bezerra \email{diogo.bezerra@ufpe.br}
}

}
\keyword{internal}
