% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comexr-package.R
\docType{package}
\name{comexr-package}
\alias{comexr}
\alias{comexr-package}
\title{comexr: Client for the Brazilian Foreign Trade Statistics API ('ComexStat')}
\description{
Interface to the 'ComexStat' API \url{https://comexstat.mdic.gov.br/} from the Brazilian Ministry of Development, Industry, Trade and Services (MDIC). Provides access to detailed export and import data, including general trade statistics (1997-present), city-level data, historical data (1989-1996), and auxiliary tables with product codes (NCM - Nomenclatura Comum do Mercosul, NBM - Nomenclatura Brasileira de Mercadorias, HS - Harmonized System), countries, economic classifications (CGCE - Classificacao por Grandes Categorias Economicas, SITC - Standard International Trade Classification, ISIC - International Standard Industrial Classification), and other categories. Uses only 'httr2' for HTTP requests and 'cli' for console messages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://strategicprojects.github.io/comexr/}
  \item \url{https://github.com/StrategicProjects/comexr}
  \item Report bugs at \url{https://github.com/StrategicProjects/comexr/issues}
}

}
\author{
\strong{Maintainer}: Andre Leite \email{leite@castlab.org}

Authors:
\itemize{
  \item Marcos Wasilew \email{marcos.wasilew@gmail.com}
  \item Hugo Vasconcelos \email{hugo.vasconcelos@ufpe.br}
  \item Carlos Amorin \email{carlos.agaf@ufpe.br}
  \item Diogo Bezerra \email{diogo.bezerra@ufpe.br}
}

}
\keyword{internal}
