% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{comexstat}
\alias{comexstat}
\title{Reads comexstat data from the data directory}
\usage{
comexstat(table = "trade", ...)
}
\value{
tibble available columns are
co_ano: aa,
co_mes, co_ncm, fluxo, co_pais, sg_uf_ncm, co_via, co_urf, co_unid, qt_estat, kg_liquido, vl_fob, vl_frete, vl_seguro, vl_cif
}
\description{
Reads comexstat data from the data directory
}
\details{
After successfully running comexstat_download(), comexstat data will be in the data directory. This function reads the data using arrow, allowing fast read, particularly when reading subsets of the data.

It defaults to the trade data (imports and exports values) using arrow. For this case you might have to use collect() to actually read the data.
Note that arrow only accepts a subset of the dplyr functions. The remaining tables are read from the original csv files.
}
\examples{
comexstat_download()
comexstat()|>filter(co_ano>2017)|>group_by(co_ano, fluxo)|>summarise(vl_fob_bi=sum(vl_fob)/1e9)|>arrange(co_ano)|>collect
}
