% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{comexstat_deflated}
\alias{comexstat_deflated}
\title{Deflate comexstat series}
\usage{
comexstat_deflated(
  data = comexstat(),
  basedate = NULL,
  deflators = get_deflators(na_omit = TRUE)
)
}
\arguments{
\item{data}{trade data to deflate}

\item{basedate}{base date to deflate to.}

\item{deflators}{data.frame with deflators}
}
\value{
deflated data.
}
\description{
Deflate comexstat series
}
