% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{comexstat_download}
\alias{comexstat_download}
\title{Download raw data from Comexstat}
\usage{
comexstat_download(..., force_download = FALSE, increase_timeout = TRUE)
}
\arguments{
\item{...}{arguments sent to the download.file function}

\item{force_download}{Force downloading data, even if it is (apparently) current.}
}
\description{
Download raw data from Comexstat
}
\details{
Depending on local cofiguration and network properties, you might need to set the download methods and other parameters, which are sent to the download.file function.
}
\examples{
\dontrun{
comexstat_download()
## might need something like this if you get ssl errors.
comexstat_download(method="wget", extra="--no-check-certificate")
## or this, if the default method does not work
comexstat_download(method="libcurl")
## or this, if it times out
options(timeout=100)
}
}
