% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{comexstat_download}
\alias{comexstat_download}
\title{Download raw data from Comexstat}
\usage{
comexstat_download(force_download = FALSE, rewrite = TRUE)
}
\arguments{
\item{force_download}{Force downloading data, even if it is (apparently) current.}
}
\description{
Download raw data from Comexstat
}
\examples{
\dontrun{ comexstat_download() }
}
