% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comexstat2.R
\name{get_deflators}
\alias{get_deflators}
\title{Download and processes deflators (CPI/USA, IPCA/Brazil, Exchange rate BRL/USD)}
\usage{
get_deflators(updated = Sys.Date(), na_omit = FALSE)
}
\arguments{
\item{updated}{}
}
\value{
A data frame with columns date, ipca (monthly inflation from Brazil), ipca_i (monthly inflation from Brazil indexed such as 1997-01-01 is 1), cpi index (monthly inflation from USA).
}
\description{
Download and processes deflators (CPI/USA, IPCA/Brazil, Exchange rate BRL/USD)
}
